Use floats in temp program

+++ temp.c
@@ -2,7 +2,7 @@
 
 int main()
 {
-  int fahr, celsius;
+  float fahr, celsius;
   int lower, upper, step;
 
   lower = 0;
@@ -12,7 +12,7 @@ int main()
   fahr = lower;
   while (fahr <= upper) {
     celsius = 5 * (fahr-32) / 9;
-    printf("%3d %6d\n", fahr, celsius);
+    printf("%3.0f %6.2f\n", fahr, celsius);
     fahr = fahr + step;
   }
 }
